fix: Hot reload of browse pages
