--- app/browser/dataset-browser.tsx
@@ -1,32 +1,0 @@
-import { AppLayout } from "@/components/layout";
-import { SelectDatasetStep } from "@/configurator/components/select-dataset-step";
-import { DataCubeResultOrder } from "@/graphql/query-hooks";
-import { EditorConfiguratorStateProvider } from "@/src";
-
-export type BrowseParams = {
-  type?: "theme" | "organization" | "dataset";
-  subtype?: "theme" | "organization";
-  iri?: string;
-  subiri?: string;
-  topic?: string;
-  search?: string;
-  order?: DataCubeResultOrder;
-  includeDrafts?: boolean;
-  dataset?: string;
-};
-
-// Generic component for all browse subpages
-export const DatasetBrowser = () => {
-  return (
-    <AppLayout>
-      <EditorConfiguratorStateProvider
-        chartId="new"
-        allowDefaultRedirect={false}
-      >
-        <SelectDatasetStep />
-      </EditorConfiguratorStateProvider>
-    </AppLayout>
-  );
-};
-
-export default DatasetBrowser;
