--- app/browser/index.tsx
@@ -1,1 +1,0 @@
-export { default as DatasetBrowser } from "./dataset-browser";
