+++ app/configurator/components/dataset-browse.spec.tsx
@@ -1,6 +1,6 @@
-import { BrowseParams } from "@/browser/dataset-browser";
 import { getFiltersFromParams } from "@/configurator/components/dataset-browse";
 import { DataCubeOrganization, DataCubeTheme } from "@/graphql/query-hooks";
+import { BrowseParams } from "@/pages/browse";
 
 const ctx = {
   themes: [
