+++ app/pages/browse/[type]/[iri]/[subtype]/[subiri].tsx
@@ -1,2 +1,3 @@
-import { DatasetBrowser } from "@/browser";
+import { DatasetBrowser } from "@/pages/browse";
+
 export default DatasetBrowser;
