+++ app/pages/browse/index.tsx
@@ -1,3 +1,31 @@
-import { DatasetBrowser } from "@/browser";
+import { AppLayout } from "@/components/layout";
+import { SelectDatasetStep } from "@/configurator/components/select-dataset-step";
+import { EditorConfiguratorStateProvider } from "@/configurator/configurator-state";
+import { DataCubeResultOrder } from "@/graphql/query-hooks";
+
+export type BrowseParams = {
+  type?: "theme" | "organization" | "dataset";
+  subtype?: "theme" | "organization";
+  iri?: string;
+  subiri?: string;
+  topic?: string;
+  search?: string;
+  order?: DataCubeResultOrder;
+  includeDrafts?: boolean;
+  dataset?: string;
+};
+
+export function DatasetBrowser() {
+  return (
+    <AppLayout>
+      <EditorConfiguratorStateProvider
+        chartId="new"
+        allowDefaultRedirect={false}
+      >
+        <SelectDatasetStep />
+      </EditorConfiguratorStateProvider>
+    </AppLayout>
+  );
+}
 
 export default DatasetBrowser;
